fix: evtrigs ownership
Fixes https://github.com/supabase/postgres/issues/1437.

Now that the effects of the `migrations/db/init-scripts` are fully
tested on the following files:

- [nix/tests/sql/auth.out](https://github.com/supabase/postgres/blob/develop/nix/tests/expected/auth.out)
- [nix/tests/sql/storage.out](https://github.com/supabase/postgres/blob/develop/nix/tests/expected/storage.out)
- [nix/tests/sql/roles.out](https://github.com/supabase/postgres/blob/develop/nix/tests/expected/roles.out)
- [nix/tests/sql/evtrigs.out](https://github.com/supabase/postgres/blob/develop/nix/tests/expected/evtrigs.out)
- [nix/tests/sql/extensions_schema.out](https://github.com/supabase/postgres/blob/develop/nix/tests/expected/extensions_schema.out)
- [nix/tests/sql/realtime.out](https://github.com/supabase/postgres/blob/develop/nix/tests/expected/realtime.out)

This reattempts to move `migrations/db/init-scripts` to `migrations/db/migrations` and have a single directory for migrations.

Of the above tests, the only change pertains to the event triggers owners which is the problem on https://github.com/supabase/postgres/issues/1437.

This requires modifying migrations, but the changes are more about following best practices:

+ Makes migrations idempotent.
+ [alter default privileges](https://www.postgresql.org/docs/current/sql-alterdefaultprivileges.html) explicitly list postgres as the target role.
+ Also:
  - adds a conditional for pgjwt which is not available on the 17 version
  - Realtime publication maintains the same postgres owner